# swagger/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"ConversationType": {
				"enum": [
					"AI_CHAT",
					"DIRECT",
					"GROUP"
				],
				"type": "string"
			},
			"MessageType": {
				"enum": [
					"TEXT",
					"IMAGE",
					"FILE"
				],
				"type": "string"
			},
			"SenderType": {
				"enum": [
					"SYSTEM",
					"AI",
					"USER"
				],
				"type": "string"
			},
			"ConversationDTO": {
				"properties": {
					"id": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"type": {
						"$ref": "#/components/schemas/ConversationType"
					},
					"userId": {
						"type": "string"
					},
					"ownerId": {
						"type": "string"
					},
					"messages": {
						"items": {
							"$ref": "#/components/schemas/MessageDTO"
						},
						"type": "array"
					},
					"members": {
						"items": {
							"$ref": "#/components/schemas/ConversationMember"
						},
						"type": "array"
					},
					"read": {
						"type": "boolean"
					}
				},
				"required": [
					"id",
					"type",
					"userId",
					"ownerId",
					"messages",
					"members",
					"read"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserDTO": {
				"properties": {
					"id": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					},
					"firstName": {
						"type": "string"
					},
					"createdAt": {
						"type": "string"
					},
					"phoneNumber": {
						"type": "string"
					},
					"birthDate": {
						"type": "string",
						"nullable": true
					},
					"active": {
						"type": "boolean"
					},
					"conversations": {
						"items": {
							"$ref": "#/components/schemas/ConversationDTO"
						},
						"type": "array"
					},
					"messaages": {
						"items": {
							"$ref": "#/components/schemas/MessageDTO"
						},
						"type": "array"
					},
					"pdpUrl": {
						"type": "string"
					},
					"roleId": {
						"type": "string"
					}
				},
				"required": [
					"email",
					"lastName",
					"firstName",
					"phoneNumber",
					"roleId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"MessageDTO": {
				"properties": {
					"content": {
						"type": "string"
					},
					"type": {
						"$ref": "#/components/schemas/MessageType"
					},
					"sender": {
						"$ref": "#/components/schemas/SenderType"
					},
					"conversation": {
						"$ref": "#/components/schemas/ConversationDTO"
					},
					"user": {
						"$ref": "#/components/schemas/UserDTO"
					}
				},
				"required": [
					"content",
					"type",
					"sender",
					"conversation",
					"user"
				],
				"type": "object",
				"additionalProperties": false
			},
			"MemberRole": {
				"enum": [
					"ADMIN",
					"MEMBER"
				],
				"type": "string"
			},
			"ConversationMember": {
				"properties": {
					"id": {
						"type": "string"
					},
					"conversationId": {
						"type": "string"
					},
					"userId": {
						"type": "string"
					},
					"role": {
						"$ref": "#/components/schemas/MemberRole"
					},
					"joinedAt": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"conversationId",
					"userId",
					"role",
					"joinedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"SprintStatus": {
				"enum": [
					"PLANNED",
					"ACTIVE",
					"CLOSED"
				],
				"type": "string"
			},
			"SprintDTO": {
				"properties": {
					"id": {
						"type": "string"
					},
					"projectId": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"goal": {
						"type": "string",
						"nullable": true
					},
					"startDate": {
						"type": "string",
						"nullable": true
					},
					"endDate": {
						"type": "string",
						"nullable": true
					},
					"status": {
						"$ref": "#/components/schemas/SprintStatus"
					},
					"createdAt": {
						"type": "string"
					},
					"updatedAt": {
						"type": "string"
					}
				},
				"required": [
					"projectId",
					"name",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateSprintRequestDTO": {
				"properties": {
					"projectId": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"goal": {
						"type": "string"
					},
					"startDate": {
						"type": "string"
					},
					"endDate": {
						"type": "string"
					}
				},
				"required": [
					"projectId",
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateSprintRequestDTO": {
				"properties": {
					"name": {
						"type": "string"
					},
					"goal": {
						"type": "string",
						"nullable": true
					},
					"startDate": {
						"type": "string",
						"nullable": true
					},
					"endDate": {
						"type": "string",
						"nullable": true
					},
					"status": {
						"$ref": "#/components/schemas/SprintStatus"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"Partial_UserDTO_": {
				"properties": {
					"id": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					},
					"firstName": {
						"type": "string"
					},
					"createdAt": {
						"type": "string"
					},
					"phoneNumber": {
						"type": "string"
					},
					"birthDate": {
						"type": "string"
					},
					"active": {
						"type": "boolean"
					},
					"conversations": {
						"items": {
							"$ref": "#/components/schemas/ConversationDTO"
						},
						"type": "array"
					},
					"messaages": {
						"items": {
							"$ref": "#/components/schemas/MessageDTO"
						},
						"type": "array"
					},
					"pdpUrl": {
						"type": "string"
					},
					"roleId": {
						"type": "string"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
...
			"IssueType": {
				"enum": [
					"STORY",
					"TASK",
					"BUG",
					"EPIC"
				],
				"type": "string"
			},
			"IssueStatus": {
				"enum": [
					"TODO",
					"IN_PROGRESS",
					"IN_REVIEW",
					"DONE",
					"CANCELLED"
				],
				"type": "string"
			},
			"IssuePriority": {
				"enum": [
					"LOW",
					"MEDIUM",
					"HIGH",
					"CRITICAL"
				],
				"type": "string"
			},
			"IssueLabelDTO": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"color": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"name",
					"color"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IssueDTO": {
				"properties": {
					"id": {
						"type": "string"
					},
					"projectId": {
						"type": "string"
					},
					"number": {
						"type": "number",
						"format": "double"
					},
					"key": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string",
						"nullable": true
					},
					"type": {
						"$ref": "#/components/schemas/IssueType"
					},
					"status": {
						"$ref": "#/components/schemas/IssueStatus"
					},
					"priority": {
						"$ref": "#/components/schemas/IssuePriority"
					},
					"storyPoints": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"position": {
						"type": "number",
						"format": "double"
					},
					"sprintId": {
						"type": "string",
						"nullable": true
					},
					"assigneeId": {
						"type": "string",
						"nullable": true
					},
					"assignee": {
						"$ref": "#/components/schemas/Partial_UserDTO_"
					},
					"reporterId": {
						"type": "string"
					},
					"reporter": {
						"$ref": "#/components/schemas/Partial_UserDTO_"
					},
					"parentIssueId": {
						"type": "string",
						"nullable": true
					},
					"labels": {
						"items": {
							"$ref": "#/components/schemas/IssueLabelDTO"
						},
						"type": "array"
					},
					"createdAt": {
						"type": "string"
					},
					"updatedAt": {
						"type": "string"
					}
				},
				"required": [
					"projectId",
					"title",
					"type",
					"status",
					"priority",
					"reporterId"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"CreateLabelRequestDTO": {
				"properties": {
					"projectId": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"color": {
						"type": "string"
					}
				},
				"required": [
					"projectId",
					"name",
					"color"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateIssueRequestDTO": {
				"properties": {
					"projectId": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"type": {
						"$ref": "#/components/schemas/IssueType"
					},
					"priority": {
						"$ref": "#/components/schemas/IssuePriority"
					},
					"storyPoints": {
						"type": "number",
						"format": "double"
					},
					"sprintId": {
						"type": "string"
					},
					"assigneeId": {
						"type": "string"
					},
					"parentIssueId": {
						"type": "string"
					},
					"labelIds": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"projectId",
					"title"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateIssueRequestDTO": {
				"properties": {
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string",
						"nullable": true
					},
					"type": {
						"$ref": "#/components/schemas/IssueType"
					},
					"status": {
						"$ref": "#/components/schemas/IssueStatus"
					},
					"priority": {
						"$ref": "#/components/schemas/IssuePriority"
					},
					"storyPoints": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"position": {
						"type": "number",
						"format": "double"
					},
					"sprintId": {
						"type": "string",
						"nullable": true
					},
					"assigneeId": {
						"type": "string",
						"nullable": true
					},
					"parentIssueId": {
						"type": "string",
						"nullable": true
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"IssueCommentDTO": {
				"properties": {
					"id": {
						"type": "string"
					},
					"content": {
						"type": "string"
					},
					"issueId": {
						"type": "string"
					},
					"authorId": {
						"type": "string"
					},
					"author": {
						"$ref": "#/components/schemas/Partial_UserDTO_"
					},
					"parentId": {
						"type": "string",
						"nullable": true
					},
					"createdAt": {
						"type": "string"
					},
					"updatedAt": {
						"type": "string"
					}
				},
				"required": [
					"content",
					"issueId",
					"authorId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateIssueCommentRequestDTO": {
				"properties": {
					"issueId": {
						"type": "string"
					},
					"content": {
						"type": "string"
					},
					"parentId": {
						"type": "string"
					}
				},
				"required": [
					"issueId",
					"content"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"bearer": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "esaina_back",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
...
		"/sprint/create": {
			"post": {
				"operationId": "Create",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"data": {
											"$ref": "#/components/schemas/SprintDTO"
										},
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"data",
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"sprint"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateSprintRequestDTO"
							}
						}
					}
				}
			}
		},
		"/sprint/list": {
			"get": {
				"operationId": "List",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"data": {
											"items": {
												"$ref": "#/components/schemas/SprintDTO"
											},
											"type": "array"
										},
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"data",
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"sprint"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "projectId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/sprint/{sprintId}": {
			"patch": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"data": {
											"$ref": "#/components/schemas/SprintDTO"
										},
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"data",
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"sprint"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "sprintId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateSprintRequestDTO"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "Remove",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"message",
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"sprint"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "sprintId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/sprint/{sprintId}/start": {
			"post": {
				"operationId": "Start",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"data": {
											"$ref": "#/components/schemas/SprintDTO"
										},
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"data",
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"sprint"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "sprintId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/sprint/{sprintId}/close": {
			"post": {
				"operationId": "Close",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"data": {
											"$ref": "#/components/schemas/SprintDTO"
										},
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"data",
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"sprint"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "sprintId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/label/create": {
			"post": {
				"operationId": "Create",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"data": {
											"properties": {
												"color": {
													"type": "string"
												},
												"name": {
													"type": "string"
												},
												"id": {
													"type": "string"
												}
											},
											"required": [
												"color",
												"name",
												"id"
											],
											"type": "object"
										},
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"data",
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"label"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateLabelRequestDTO"
							}
						}
					}
				}
			}
		},
		"/label/list": {
			"get": {
				"operationId": "List",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"data": {
											"items": {
												"properties": {
													"color": {
														"type": "string"
													},
													"name": {
														"type": "string"
													},
													"id": {
														"type": "string"
													}
												},
												"required": [
													"color",
													"name",
													"id"
												],
												"type": "object"
											},
											"type": "array"
										},
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"data",
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"label"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "projectId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/label/{labelId}": {
			"delete": {
				"operationId": "Remove",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"message",
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"label"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "labelId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/issue/create": {
			"post": {
				"operationId": "Create",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"data": {
											"$ref": "#/components/schemas/IssueDTO"
										},
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"data",
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"issue"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateIssueRequestDTO"
							}
						}
					}
				}
			}
		},
		"/issue/list": {
			"get": {
				"operationId": "List",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"data": {
											"items": {
												"$ref": "#/components/schemas/IssueDTO"
											},
											"type": "array"
										},
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"data",
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"issue"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "projectId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "sprintId",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "assigneeId",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "status",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/IssueStatus"
						}
					},
					{
						"in": "query",
						"name": "type",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/IssueType"
						}
					}
				]
			}
		},
		"/issue/{issueId}": {
			"get": {
				"operationId": "GetById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"data": {
											"$ref": "#/components/schemas/IssueDTO"
										},
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"data",
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"issue"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "issueId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"patch": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"data": {
											"$ref": "#/components/schemas/IssueDTO"
										},
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"data",
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"issue"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "issueId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateIssueRequestDTO"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "Remove",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"message",
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"issue"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "issueId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/issue/comment": {
			"post": {
				"operationId": "AddComment",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"data": {
											"$ref": "#/components/schemas/IssueCommentDTO"
										},
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"data",
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"issue"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateIssueCommentRequestDTO"
							}
						}
					}
				}
			}
		},
		"/issue/{issueId}/comments": {
			"get": {
				"operationId": "ListComments",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"data": {
											"items": {
												"$ref": "#/components/schemas/IssueCommentDTO"
											},
											"type": "array"
										},
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"data",
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"issue"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "issueId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}